requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.0
sqlalchemy==2.0.45
alembic==1.17.2
psycopg2-binary==2.9.11
email-validator==2.3.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.21